move error handling functions from esm_parser to esm_tools (into error_handling.py module). Black error_handling.py. isort all .py files containing changes
