+++ src/esm_runscripts/cli.py
@@ -15,7 +15,7 @@ import sys
 from loguru import logger
 
 from esm_motd import check_all_esm_packages
-from esm_parser import user_error
+from esm_tools import user_error
 
 from .helpers import SmartSink
 from .sim_objects import *
