+++ src/esm_runscripts/config_initialization.py
@@ -4,6 +4,7 @@ import sys
 
 import esm_parser
 import esm_tools
+from esm_tools import user_error
 
 from . import chunky_parts
 
@@ -161,7 +162,7 @@ def get_user_config_from_command_line(command_line_config):
     except SystemExit as sysexit:
         sys.exit(sysexit)
     except:
-        esm_parser.user_error(
+        user_error(
             "Syntax error",
             f"An error occurred while reading the config file "
             f"``{command_line_config['runscript_abspath']}`` from the command line.",
@@ -284,7 +285,7 @@ def check_account(config):
     # Check if the 'account' variable is needed and missing
     if config["computer"].get("accounting", False):
         if "account" not in config["general"]:
-            esm_parser.user_error(
+            user_error(
                 "Missing account info",
                 f"You cannot run simulations in '{config['computer']['name']}' "
                 "without providing an 'account' variable in the 'general' section, whose "
