+++ src/esm_runscripts/oasis.py
@@ -4,10 +4,9 @@ import subprocess
 import sys
 
 import questionary
-
-from esm_parser import user_error
 from loguru import logger
 
+from esm_tools import user_error
 
 
 class oasis:
