+++ src/esm_runscripts/prev_run.py
@@ -3,10 +3,11 @@ import sys
 
 import questionary
 import yaml
+from loguru import logger
 
 import esm_parser
 from esm_calendar import Calendar, Date
-from loguru import logger
+from esm_tools import user_error, user_note
 
 
 class PrevRunInfo(dict):
@@ -241,7 +242,7 @@ class PrevRunInfo(dict):
                 )
                 # Dates don't match
                 if calc_prev_date_stamp != prev_date_stamp and self.warn:
-                    esm_parser.user_note(
+                    user_note(
                         f"End date of the previous configuration file for '{component}'"
                         + " not coinciding:",
                         (
@@ -330,7 +331,7 @@ class PrevRunInfo(dict):
                     "prev_run_config_file"
                 )
                 if not user_prev_run_config_full_path:
-                    esm_parser.user_error(
+                    user_error(
                         "'prev_run_config_file' not defined",
                         "You are trying to run a branchoff experiment that uses the "
                         + f"'prev_run' functionality for '{component}' without "
@@ -368,7 +369,7 @@ class PrevRunInfo(dict):
 
         # Check for errors
         if not os.path.isdir(config_dir):
-            esm_parser.user_error(
+            user_error(
                 "Config folder not existing",
                 (
                     f"The config folder {config_dir} does not exist. "
@@ -430,7 +431,7 @@ class PrevRunInfo(dict):
                     )
                 else:
                     # Error
-                    esm_parser.user_error(
+                    user_error(
                         "Too many possible config files",
                         "There is more than one config file with the same final date "
                         + "as the one required for the continuation of this experiment."
@@ -450,7 +451,7 @@ class PrevRunInfo(dict):
             prev_run_config_file = user_prev_run_config_file
             prev_run_config_path = f"{config_dir}/{prev_run_config_file}"
             if not os.path.isfile(prev_run_config_path):
-                esm_parser.user_error(
+                user_error(
                     "'prev_run_config_file' incorrectly defined",
                     f"The file defined in the '{component}.prev_run_config_path' "
                     + f"({prev_run_config_path}) does not exist.",
