+++ src/esm_tools/__init__.py
@@ -38,6 +38,8 @@ import pkg_resources
 import yaml
 from loguru import logger
 
+from .error_handling import *
+
 # Setup Loguru for the following cases:
 # A) If user sets
 if os.environ.get("DEBUG_ESM_TOOLS"):
